hack: for prev_page, ensure the returned page actually ends at the start of the previous page.
With the multifile implementation, the request "go back a page size" can result in it
loading less then a page size of data from disk- primarily due to a file being smaller than
the page size.  When this occurs, it results in the returned page not being the literal
page from prior to the current one, breaking the API contract for consuming code.

As such, add a check in the core cfile implementation; longer term, this should be made
smarter/faster.