tree format: Increase efficiency via building a UID/GID/Mode table, and referencing that.
Roughly; the uid/gid/mode's were 11 bytes of data; however, there are only a few
combinations in reality.  As such, exploit that and use a variable length table
of those tuple combos, and for each command, write the index position instead.

Additionally, store just the permission bits of mode; this increases UGM efficiency,
and ensures we're not storing crap data we shouldn't be (such as S_IFSOCK).  Finally
the remaining mode data is < 16 bits thus limit mode to 2 bytes.