Trade ~1% delta increase for ~38% runtime reduction.
Previously, it would wind up frequently sampling every single byte.
In principle, that sounds like a non issue; in reality, it results in
rarely an actual extent match that wouldn't be found w/ a slightly
higher sample rate.  If the seed is 512 (requiring a match of 512 bytes),
sampling every single byte is pointless.

Adjusting the logic so it uses the maximum of a 1/16th of the seed
size, or every 2 bytes (for seed's <=32).  Doing a simple differ between
two portage snapshots of ~100MB each, this drops run time from 126s to
78s, w/ ~1% increase (reduced to ~.1% increase after compression).

Workable tradeoff, in other words.