Increase default allowed memory usage for better compression.
The reality is that the memory usage doesn't ever hit the max due to
failures in the hashing approach; however, these changes *do* ensure
there is a better utilization of the hash w/out impacting memory
utilization for the most part.

Effectively, for 2 100MB files, we utilize a <300MB of ram at max,
but get a 1.5% improvement- both uncompressed and compressed- due to
the increased matching.

It's a workaround for flaws in the rh bucket implementation, but it's
usable and worth the addition (especially since hash size is configurable
via commandline).