Drop dead/unused hash implementations.
These are all old implementations that were used; rather than just
removing them, they were left available for usage.

They're not useful in this day age for the most part; available memory
is high enough these sort of one off hacks aren't worth it- nor have
those code paths been executed in a long while.

As such, drop the dead code; this is a step towards moving the hashing
implemention so it's an internal detail- allowing for inlining/optimization,
and simplifying the implementations.