Wipe old GPL reference, add SPDX bsd-3 identifier for file copyright.
Diffball was relicensed to bsd-3 years ago (with the central 'COPYING' file
updated to reflect that), but the GPL headers weren't scrubbed as they should
have been.

This commit does that, whilst also adding SPDX identifiers, and updating the
copyright range to be final/accurate (derived from git).

This is a cleanup commit for anyone who uses the code in the future; this
isn't intended as me restarting development on diffball.