Simplify multipass logic while maintaining the same byte output.
Comments were added to clarify, but multipass has a behaviour where it
inverts what's being hashed, using that to try and minimize hash collisions.

The utility of that logic should be revisited, but in the meantime, the two
branches of code are effectively the same sans the hash functions, thus unify
it down to one common codepath.

Signed-off-by: Brian Harring <ferringb@gmail.com>